Rename test files and update the prerequisites
